DjangoxAi/main/requirements.txt
Django>=5.0,<6.0
jupyter
langchain-cohere
langgraph
langgraph-supervisor
python-decouple
permit
djangorestframework